+++ app/lib/tool/neat_task/pub_dev_tasks.dart
@@ -30,7 +30,6 @@ import '../../task/global_lock.dart';
 import '../../tool/backfill/backfill_new_fields.dart';
 import '../maintenance/remove_orphaned_likes.dart';
 import '../maintenance/update_package_likes.dart';
-import '../maintenance/update_public_bucket.dart';
 import 'datastore_status_provider.dart';
 
 final _logger = Logger('pub_dev_tasks');
@@ -114,7 +113,8 @@ void _setupGenericPeriodicTasks() {
   _daily(
     name: 'sync-public-bucket-from-canonical-bucket',
     isRuntimeVersioned: false,
-    task: updatePublicArchiveBucket,
+    task: () async =>
+        await packageBackend.packageStorage.updatePublicArchiveBucket(),
   );
 
   // Exports the package name completion data to a bucket.
