+++ app/test/tool/maintenance/update_public_bucket_test.dart
@@ -3,17 +3,29 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:gcloud/storage.dart';
+import 'package:pub_dev/package/backend.dart';
+import 'package:pub_dev/package/package_storage.dart';
 import 'package:pub_dev/shared/configuration.dart';
-import 'package:pub_dev/tool/maintenance/update_public_bucket.dart';
 import 'package:test/test.dart';
 
 import '../../shared/test_services.dart';
 
 void main() {
   group('update public bucket', () {
+    Future<PublicBucketUpdateStat> _updatePublicArchiveBucket({
+      String? package,
+      Duration? ageCheckThreshold,
+      Duration? deleteIfOlder,
+    }) async {
+      return await packageBackend.packageStorage.updatePublicArchiveBucket(
+        package: package,
+        ageCheckThreshold: ageCheckThreshold ?? Duration.zero,
+        deleteIfOlder: deleteIfOlder ?? const Duration(days: 7),
+      );
+    }
+
     testWithProfile('no update', fn: () async {
-      final changes =
-          await updatePublicArchiveBucket(ageCheckThreshold: Duration.zero);
+      final changes = await _updatePublicArchiveBucket();
       expect(changes.isAllZero, isTrue);
     });
 
@@ -22,14 +34,12 @@ void main() {
           storageService.bucket(activeConfiguration.publicPackagesBucketName!);
       await bucket.delete('packages/oxygen-1.0.0.tar.gz');
 
-      final changes =
-          await updatePublicArchiveBucket(ageCheckThreshold: Duration.zero);
+      final changes = await _updatePublicArchiveBucket();
       expect(changes.archivesUpdated, 1);
       expect(changes.archivesToBeDeleted, 0);
       expect(changes.archivesDeleted, 0);
 
-      final changes2 =
-          await updatePublicArchiveBucket(ageCheckThreshold: Duration.zero);
+      final changes2 = await _updatePublicArchiveBucket();
       expect(changes2.isAllZero, isTrue);
     });
 
@@ -39,22 +49,15 @@ void main() {
           storageService.bucket(activeConfiguration.publicPackagesBucketName!);
       await bucket.delete('packages/oxygen-1.0.0.tar.gz');
 
-      final changes = await updatePublicArchiveBucket(
-        package: 'oxygen',
-        ageCheckThreshold: Duration.zero,
-      );
+      final changes = await _updatePublicArchiveBucket(package: 'oxygen');
       expect(changes.archivesUpdated, 1);
       expect(changes.archivesToBeDeleted, 0);
       expect(changes.archivesDeleted, 0);
 
-      final changes2 = await updatePublicArchiveBucket(
-        package: 'oxygen',
-        ageCheckThreshold: Duration.zero,
-      );
+      final changes2 = await _updatePublicArchiveBucket(package: 'oxygen');
       expect(changes2.isAllZero, isTrue);
 
-      final changes3 =
-          await updatePublicArchiveBucket(ageCheckThreshold: Duration.zero);
+      final changes3 = await _updatePublicArchiveBucket();
       expect(changes3.isAllZero, isTrue);
     });
 
@@ -64,21 +67,14 @@ void main() {
           storageService.bucket(activeConfiguration.publicPackagesBucketName!);
       await bucket.delete('packages/oxygen-1.0.0.tar.gz');
 
-      final changes = await updatePublicArchiveBucket(
-        package: 'neon',
-        ageCheckThreshold: Duration.zero,
-      );
+      final changes = await _updatePublicArchiveBucket(package: 'neon');
       expect(changes.isAllZero, isTrue);
 
-      final changes2 = await updatePublicArchiveBucket(
-        package: 'neon',
-        ageCheckThreshold: Duration.zero,
-      );
+      final changes2 = await _updatePublicArchiveBucket(package: 'neon');
       expect(changes2.isAllZero, isTrue);
 
       // eventual update
-      final changes3 =
-          await updatePublicArchiveBucket(ageCheckThreshold: Duration.zero);
+      final changes3 = await _updatePublicArchiveBucket();
       expect(changes3.archivesUpdated, 1);
       expect(changes3.archivesToBeDeleted, 0);
       expect(changes3.archivesDeleted, 0);
@@ -90,19 +86,18 @@ void main() {
       await bucket.writeBytes('packages/oxygen-0.0.99.tar.gz', [1]);
 
       // recent file gets ignored
-      final recent = await updatePublicArchiveBucket();
+      final recent = await _updatePublicArchiveBucket(
+          ageCheckThreshold: Duration(days: 1));
       expect(recent.isAllZero, isTrue);
 
       // non-recent file will get deleted
-      final changes =
-          await updatePublicArchiveBucket(ageCheckThreshold: Duration.zero);
+      final changes = await _updatePublicArchiveBucket();
       expect(changes.archivesUpdated, 0);
       expect(changes.archivesToBeDeleted, 1);
       expect(changes.archivesDeleted, 0);
 
       // non-recent file is deleted
-      final changes2 = await updatePublicArchiveBucket(
-        ageCheckThreshold: Duration.zero,
+      final changes2 = await _updatePublicArchiveBucket(
         deleteIfOlder: Duration.zero,
       );
       expect(changes2.archivesUpdated, 0);
@@ -110,8 +105,7 @@ void main() {
       expect(changes2.archivesDeleted, 1);
 
       // second round should report 0 deleted
-      final changes3 =
-          await updatePublicArchiveBucket(ageCheckThreshold: Duration.zero);
+      final changes3 = await _updatePublicArchiveBucket();
       expect(changes3.isAllZero, isTrue);
     });
 
@@ -121,22 +115,19 @@ void main() {
       await bucket.writeBytes('packages/oxygen-0.0.99.tar.gz', [1]);
 
       // recent file gets ignored
-      final recent = await updatePublicArchiveBucket(package: 'oxygen');
+      final recent = await _updatePublicArchiveBucket(
+          ageCheckThreshold: Duration(days: 1));
       expect(recent.isAllZero, isTrue);
 
       // non-recent file will get deleted
-      final changes = await updatePublicArchiveBucket(
-        package: 'oxygen',
-        ageCheckThreshold: Duration.zero,
-      );
+      final changes = await _updatePublicArchiveBucket(package: 'oxygen');
       expect(changes.archivesUpdated, 0);
       expect(changes.archivesToBeDeleted, 1);
       expect(changes.archivesDeleted, 0);
 
       // non-recent file is deleted
-      final changes2 = await updatePublicArchiveBucket(
+      final changes2 = await _updatePublicArchiveBucket(
         package: 'oxygen',
-        ageCheckThreshold: Duration.zero,
         deleteIfOlder: Duration.zero,
       );
       expect(changes2.archivesUpdated, 0);
@@ -144,15 +135,11 @@ void main() {
       expect(changes2.archivesDeleted, 1);
 
       // second round should report 0 deleted
-      final changes3 = await updatePublicArchiveBucket(
-        package: 'oxygen',
-        ageCheckThreshold: Duration.zero,
-      );
+      final changes3 = await _updatePublicArchiveBucket(package: 'oxygen');
       expect(changes3.isAllZero, isTrue);
 
       // no further changes
-      final changes4 =
-          await updatePublicArchiveBucket(ageCheckThreshold: Duration.zero);
+      final changes4 = await _updatePublicArchiveBucket();
       expect(changes4.isAllZero, isTrue);
     });
 
@@ -162,30 +149,22 @@ void main() {
       await bucket.writeBytes('packages/oxygen-0.0.99.tar.gz', [1]);
 
       // no matching file
-      final recent = await updatePublicArchiveBucket(package: 'neon');
+      final recent = await _updatePublicArchiveBucket(package: 'neon');
       expect(recent.isAllZero, isTrue);
 
       // no matching files
-      final changes = await updatePublicArchiveBucket(
-        package: 'neon',
-        ageCheckThreshold: Duration.zero,
-      );
+      final changes = await _updatePublicArchiveBucket(package: 'neon');
       expect(changes.isAllZero, isTrue);
-      final changes2 = await updatePublicArchiveBucket(
+      final changes2 = await _updatePublicArchiveBucket(
         package: 'neon',
-        ageCheckThreshold: Duration.zero,
         deleteIfOlder: Duration.zero,
       );
       expect(changes2.isAllZero, isTrue);
-      final changes3 = await updatePublicArchiveBucket(
-        package: 'neon',
-        ageCheckThreshold: Duration.zero,
-      );
+      final changes3 = await _updatePublicArchiveBucket(package: 'neon');
       expect(changes3.isAllZero, isTrue);
 
       // changes will be picked up without filter
-      final changes4 =
-          await updatePublicArchiveBucket(ageCheckThreshold: Duration.zero);
+      final changes4 = await _updatePublicArchiveBucket();
       expect(changes4.isAllZero, isFalse);
     });
   });
